+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## v0.2.0 (TBD)
+
+- Add new group of commands `equity.discovery.*()`
+- Add new command `derivatives.futures.search()`
+- Update docs related to these two new features
+- Add unit tests for this new features
+
 ## v0.1.0 (2024-03-08)
 
 - Finish development of the first version of the VietFin package.